web-scraping/requirements.txt
praw>=7.7
yfinance>=0.2.36
requests>=2.31
beautifulsoup4>=4.12
transformers>=4.36
torch>=2.1
python-dotenv>=1.0
